refactor: исправит оптимизацию places-list, добавит favorites slice
