+++ project/src/components/places-list/places-list.tsx
@@ -1,6 +1,6 @@
 import CardPlace from '../card-place/card-place';
 import { Offer } from '../../types/offer';
-import { memo, useCallback } from 'react';
+import { memo } from 'react';
 
 interface PlaceListProps {
   offers: Offer[];
@@ -9,19 +9,13 @@ interface PlaceListProps {
 }
 
 function PlacesList({ offers, className, onPlaceCardHover }: PlaceListProps): JSX.Element {
-  const handleCardActive = useCallback(
-    (valueId: number | null) => (onPlaceCardHover !== undefined ? onPlaceCardHover(valueId) : null),
-    [onPlaceCardHover],
-  );
-
   return (
     <div className={`places__list ${className}`}>
       {offers.map((offer) => (
-        <CardPlace onCardHover={handleCardActive} offer={offer} key={offer.id} />
+        <CardPlace onCardHover={onPlaceCardHover} offer={offer} key={offer.id} />
       ))}
     </div>
   );
 }
 
-// export default PlacesList;
 export default memo(PlacesList);
