+++ project/src/store/offers-slice/offers-slice.ts
@@ -1,11 +1,13 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createAsyncThunk, createSelector, createSlice } from '@reduxjs/toolkit';
 
-import { fetchFavoritesAction, fetchOfferAction, fetchOffersAction, fetchOffersNearbyAction } from '../api-actions';
-
-import { FetchStatus, NameSpace } from '../../utils/const';
+import { APIRoute, FetchStatus, NameSpace } from '../../utils/const';
 
 import { Offer } from '../../types/offer';
 import { State } from '../../types/state';
+import { selectCity, selectSortType } from '../app-slice/app-slice';
+import { sortOffers } from '../../utils/utils';
+import { handleError } from '../../services/handle-error';
+import { api } from '../index';
 
 interface InitialState {
   offers: Offer[];
@@ -19,10 +21,6 @@ interface InitialState {
   offersNearby: Offer[] | undefined;
   offersNearbyStatus: FetchStatus;
   offersNearbyError: boolean;
-
-  favoriteOffers: Offer[] | undefined;
-  favoriteOffersStatus: FetchStatus;
-  favoriteOffersError: boolean;
 }
 
 const initialState: InitialState = {
@@ -37,11 +35,37 @@ const initialState: InitialState = {
   offersNearby: [],
   offersNearbyStatus: FetchStatus.Idle,
   offersNearbyError: false,
-
-  favoriteOffers: [],
-  favoriteOffersStatus: FetchStatus.Idle,
-  favoriteOffersError: false,
 };
+
+export const fetchOffersAction = createAsyncThunk('data/fetchOffers', async () => {
+  try {
+    const { data } = await api.get<Offer[]>(APIRoute.Offers);
+    return data;
+  } catch (err) {
+    handleError(err);
+    throw err;
+  }
+});
+
+export const fetchOfferAction = createAsyncThunk('data/fetchOffer', async (id: number) => {
+  try {
+    const { data } = await api.get<Offer>(`${APIRoute.Offers}/${id}`);
+    return data;
+  } catch (err) {
+    handleError(err);
+    throw err;
+  }
+});
+
+export const fetchOffersNearbyAction = createAsyncThunk('data/fetchOffersNearby', async (id: number) => {
+  try {
+    const { data } = await api.get<Offer[]>(`${APIRoute.Offers}/${id}/nearby`);
+    return data;
+  } catch (err) {
+    handleError(err);
+    throw err;
+  }
+});
 
 export const offersSlice = createSlice({
   name: NameSpace.Offers,
@@ -81,17 +105,6 @@ export const offersSlice = createSlice({
       .addCase(fetchOffersNearbyAction.rejected, (state) => {
         state.offersNearbyStatus = FetchStatus.Failed;
         state.offersNearbyError = true;
-      })
-      .addCase(fetchFavoritesAction.pending, (state) => {
-        state.favoriteOffersStatus = FetchStatus.Pending;
-      })
-      .addCase(fetchFavoritesAction.fulfilled, (state, action) => {
-        state.favoriteOffersStatus = FetchStatus.Success;
-        state.favoriteOffers = action.payload;
-      })
-      .addCase(fetchFavoritesAction.rejected, (state) => {
-        state.favoriteOffersStatus = FetchStatus.Failed;
-        state.favoriteOffersError = true;
       });
   },
 });
@@ -104,5 +117,13 @@ export const selectOffer = (state: State) => selectOffersState(state).offer;
 export const selectOfferStatus = (state: State) => selectOffersState(state).offerStatus;
 export const selectoffersNearby = (state: State) => selectOffersState(state).offersNearby;
 export const selectoffersNearbyStatus = (state: State) => selectOffersState(state).offersNearbyStatus;
-export const selectFavoriteOffers = (state: State) => selectOffersState(state).favoriteOffers;
-export const selectFavoriteOffersStatus = (state: State) => selectOffersState(state).favoriteOffersStatus;
+
+export const selectCurrentOffers = createSelector(
+  selectCity,
+  selectSortType,
+  selectOffers,
+  (city, sortType, offers) => {
+    const filteredOffers = offers.filter((offer) => offer.city.name === city);
+    return sortOffers(sortType, filteredOffers);
+  },
+);
